Update and Delete option modification
Previously, the program does not respond if you choose update or delete when nobody has registered yet , it will just jump and ask again without displaying any message this might confuse the use and be like "what did i do wrong?"
and now i modify the code a little bit so if you choose update or delete when nobody in the registration list, it will respond and say "Nothing to be deleted , no student has registered yet!" or "Nothing to be updated, no student has registered yet!". This is really cool because if the admin try to delete or update when nobody in the list it will tell him that you can't delete or update anything because no one has been registered yet!. unintentionally he or she might enter something wrong.